Move viscous sponge into src
