+++ examples/hybrid/sphere/baroclinic_wave_utilities.jl
@@ -42,43 +42,6 @@ function rayleigh_sponge_tendency!(Yₜ, Y, p, t, colidx)
     @. Yₜ.c.uₕ[colidx] -= ᶜβ_rayleigh_uₕ[colidx] * Y.c.uₕ[colidx]
 end
 
-# Viscous sponge
-
-function viscous_sponge_cache(Y; zd_viscous = FT(15e3), κ₂ = FT(1e5))
-    ᶜz = Fields.coordinate_field(Y.c).z
-    ᶠz = Fields.coordinate_field(Y.f).z
-    ᶜαₘ = @. ifelse(ᶜz > zd_viscous, κ₂, FT(0))
-    ᶠαₘ = @. ifelse(ᶠz > zd_viscous, κ₂, FT(0))
-    zmax = maximum(ᶠz)
-    ᶜβ_viscous =
-        @. ᶜαₘ * sin(FT(π) / 2 * (ᶜz - zd_viscous) / (zmax - zd_viscous))^2
-    ᶠβ_viscous =
-        @. ᶠαₘ * sin(FT(π) / 2 * (ᶠz - zd_viscous) / (zmax - zd_viscous))^2
-    return (; ᶜβ_viscous, ᶠβ_viscous)
-end
-
-function viscous_sponge_tendency!(Yₜ, Y, p, t)
-    (; ᶜβ_viscous, ᶠβ_viscous, ᶜp) = p
-    ᶜρ = Y.c.ρ
-    ᶜuₕ = Y.c.uₕ
-    if :ρθ in propertynames(Y.c)
-        @. Yₜ.c.ρθ += ᶜβ_viscous * wdivₕ(ᶜρ * gradₕ(Y.c.ρθ / ᶜρ))
-    elseif :ρe_tot in propertynames(Y.c)
-        @. Yₜ.c.ρe_tot += ᶜβ_viscous * wdivₕ(ᶜρ * gradₕ((Y.c.ρe_tot + ᶜp) / ᶜρ))
-    elseif :ρe_int in propertynames(Y.c)
-        @. Yₜ.c.ρe_int += ᶜβ_viscous * wdivₕ(ᶜρ * gradₕ((Y.c.ρe_int + ᶜp) / ᶜρ))
-    end
-    @. Yₜ.c.uₕ +=
-        ᶜβ_viscous * (
-            wgradₕ(divₕ(ᶜuₕ)) - Geometry.project(
-                Geometry.Covariant12Axis(),
-                wcurlₕ(Geometry.project(Geometry.Covariant3Axis(), curlₕ(ᶜuₕ))),
-            )
-        )
-    @. Yₜ.f.w.components.data.:1 +=
-        ᶠβ_viscous * wdivₕ(gradₕ(Y.f.w.components.data.:1))
-end
-
 forcing_cache(Y, ::Nothing) = NamedTuple()
 
 # Held-Suarez forcing
