+++ src/ClimaAtmos.jl
@@ -11,6 +11,8 @@ import .RRTMGPInterface as RRTMGPI
 include("TurbulenceConvection/TurbulenceConvection.jl")
 import .TurbulenceConvection as TC
 
+include(joinpath("tendencies", "viscous_sponge.jl"))
+
 include("model_getters.jl") # high-level (using parsed_args) model getters
 
 end # module
